Factor out `fine_tuning`
