+++ async-openai/src/types/mod.rs
@@ -8,6 +8,7 @@ mod common;
 mod edit;
 mod file;
 mod fine_tune;
+mod fine_tuning;
 mod image;
 mod message;
 mod message_file;
@@ -24,6 +25,7 @@ pub use common::*;
 pub use edit::*;
 pub use file::*;
 pub use fine_tune::*;
+pub use fine_tuning::*;
 pub use image::*;
 pub use message::*;
 pub use message_file::*;
